# Conflicts:
#	CMakeLists.txt
#	Client/MainMenu.cpp
#	Client/main.cpp
#	Client/ray/RaylibGraphicLib.cpp
#	Engine/CMakeLists.txt
#	Engine/Graphic/IGraphicLib.h

+++ Engine/Network/Packets/PlayerMovePacket.cpp
@@ -6,7 +6,7 @@
 
 PlayerMovePacket::PlayerMovePacket() {}
 
-PlayerMovePacket::PlayerMovePacket(int playerId, const Vector2i &acceleration) : playerId(playerId),
+PlayerMovePacket::PlayerMovePacket(size_t playerId, const Vector2i &acceleration) : playerId(playerId),
                                                                                  acceleration(acceleration) {}
 
 void PlayerMovePacket::write(ByteArray &buffer) const {
